code/requirement.txt
anaconda-client==1.6.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.4.0.2
clyent==1.2.2
configparser==3.5.0
cube2sphere==0.1.8
cycler==0.10.0
decorator==4.0.10
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
ipykernel==4.5.2
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==5.2.2
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.1
MarkupSafe==0.23
matplotlib==1.5.3
mistune==0.7.3
mock==2.0.0
nb-anacondacloud==1.2.0
nb-conda==2.0.0
nb-conda-kernels==2.0.0
nbconvert==4.2.0
nbformat==4.2.0
nbpresent==3.0.2
networkx==1.11
notebook==4.3.0
numpy==1.11.2
path.py==0.0.0
pathlib2==2.2.1
pbr==1.10.0
pexpect==4.0.1
pickleshare==0.7.4
Pillow==3.4.2
prompt-toolkit==1.0.13
protobuf==3.1.0
ptyprocess==0.5.1
pycairo==1.10.0
Pygments==2.2.0
pyparsing==2.1.4
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.2.1
requests==2.12.4
scandir==1.5
scikit-image==0.12.3
scikit-learn==0.18.1
scipy==0.18.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
tensorflow-gpu==0.12.0rc1
terminado==0.6
tornado==4.4.2
traitlets==4.3.2
transforms3d==0.3
wcwidth==0.1.7
widgetsnbextension==1.2.6
